Atualizaçao com nova tela funcional

+++ app/src/main/java/com/example/metereologic_app/data/remote/RemoteDataSource.kt
@@ -5,11 +5,11 @@ import com.example.metereologic_app.data.remote.response.WeatherDataResponse
 
 interface RemoteDataSource {
 
-    suspend fun getWeatherDataResponse(lat: Float, lng: Float): WeatherDataResponse
+    suspend fun getWeatherDataResponse(lat: Double, lng: Double): WeatherDataResponse
 
     suspend fun getWeatherByCity(city: String): WeatherDataResponse
 
-    suspend fun getForecastDataResponse(lat: Float, lng: Float): ForecastDataResponse
+    suspend fun getForecastDataResponse(lat: Double, lng: Double): ForecastDataResponse
 
     suspend fun getForecastByCity(city: String): ForecastDataResponse
